adapter philips_hue: Fixed error in json
Problem was that temperature was given to the hue as float instead of as int.

+++ src/technologies/philips_hue/adapter.py
@@ -17,31 +17,31 @@ def _translate(value, leftMin, leftMax, rightMin, rightMax):
 	return int( rightMin + (valueScaled * rightSpan) )
 
 def kelvin_mired(kelvin):
-	return 1000000 / kelvin
+	return int( 1000000 / kelvin )
 
 def set_power(powerstate):
-	data = {"on":powerstate}
+	data = {"on": powerstate}
 	_send_command(data)
 
 def set_both(brightness, temperature):
 	brightness = _translate(brightness, 0.0, 1.0, 0, 254)
 	temperature = kelvin_mired(temperature)
-	data = {"bri":brightness, "ct":temperature}
+	data = {"bri": brightness, "ct": temperature}
 	_send_command(data)
 
 def set_brightness(brightness):
 	brightness = _translate(brightness, 0.0, 1.0, 0, 254)
-	data = {"bri":brightness}
+	data = {"bri": brightness}
 	_send_command(data)
 
 def set_temperature(temperature):
 	temperature = kelvin_mired(temperature)
-	data = {"ct":temperature}
+	data = {"ct": temperature}
 	_send_command(data)
 
 def _send_command(data):
 	try:
-		result = requests.put(BASEURL+"groups/0/action",data=json.dumps(data))
+		result = requests.put(BASEURL+"groups/0/action", data=json.dumps(data))
 	except (requests.exceptions.ConnectionError, TimeoutError):
 		return False
 	except Exception as e:
